Add SD card write via SPI example

+++ modules/lpc4337_m4/fatfs_ssp/inc/diskio.h
@@ -34,8 +34,8 @@ typedef enum {
 
 DSTATUS disk_initialize (BYTE);
 DSTATUS disk_status (BYTE);
-DRESULT disk_read (BYTE, BYTE*, DWORD, BYTE);
-DRESULT disk_write (BYTE, const BYTE*, DWORD, BYTE);
+DRESULT disk_read (BYTE, BYTE*, DWORD, UINT);
+DRESULT disk_write (BYTE, const BYTE*, DWORD, UINT);
 DRESULT disk_ioctl (BYTE, BYTE, void*);
